split when with client_command
